skin_lesion_classifier/requirements.txt
# =============================================================================
# Skin Lesion Classification - Dependencies
# =============================================================================

# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
pillow>=10.0.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Progress Bars
tqdm>=4.65.0

# Utilities
python-dateutil>=2.8.0

# Optional: Experiment Tracking
# wandb>=0.15.0
# mlflow>=2.5.0

# Optional: API Server (for deployment)
# fastapi>=0.100.0
# uvicorn>=0.23.0
# python-multipart>=0.0.6

